文字列の挿入関数の実装とカプセル化とそれに伴うmakefileの更新
