+++ makefile
@@ -1,4 +1,8 @@
-ziku: ziku.o 
-	gcc -o ziku ziku.o
+ziku: ziku.o shaft.o
+	gcc -o ziku ziku.o shaft.o
 ziku.o: ziku.c
 	gcc -c ziku.c
+shaft.o:shaft.c
+	gcc -c shaft.c
+clean:
+	rm -f ziku
